+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessExtension.java
@@ -234,6 +234,11 @@ public abstract class SpotlessExtension {
 		format(GherkinExtension.NAME, GherkinExtension.class, closure);
 	}
 
+	public void toml(Action<TomlExtension> closure) {
+		requireNonNull(closure);
+		format(TomlExtension.NAME, TomlExtension.class, closure);
+	}
+
 	public void go(Action<GoExtension> closure) {
 		requireNonNull(closure);
 		format(GoExtension.NAME, GoExtension.class, closure);